LayoutTests/css3/flexbox/display-flexbox-set-get.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<style>
.flexbox { display: -webkit-flexbox; }
.flexboxInline { display: -webkit-inline-flexbox; }
</style>
<script src="../../fast/js/resources/js-test-pre.js"></script>
</head>
<body>
<div class="flexbox" id="flexbox"></div>
<div class="flexboxInline" id="flexboxInline"></div>
<script>
description('Test that setting and getting display: -webkit-flexbox and -webkit-inline-flexbox works as expected');

debug("Test getting |display| set through CSS");
var flexboxElement = document.getElementById("flexbox");
shouldBe("getComputedStyle(flexboxElement, '').getPropertyValue('display')", "'-webkit-flexbox'");
var inlineFlexboxElement = document.getElementById("flexboxInline");
shouldBe("getComputedStyle(inlineFlexboxElement, '').getPropertyValue('display')", "'-webkit-inline-flexbox'");

debug("");
debug("Test getting and setting display through JS");
var element = document.createElement("div");
document.body.appendChild(element);
shouldBe("getComputedStyle(element, '').getPropertyValue('display')", "'block'");
element.style.display = "-webkit-flexbox";
shouldBe("getComputedStyle(element, '').getPropertyValue('display')", "'-webkit-flexbox'");

element = document.createElement("div");
document.body.appendChild(element);
shouldBe("getComputedStyle(element, '').getPropertyValue('display')", "'block'");
element.style.display = "-webkit-inline-flexbox";
shouldBe("getComputedStyle(element, '').getPropertyValue('display')", "'-webkit-inline-flexbox'");
</script>
<script src="../../fast/js/resources/js-test-post.js"></script>
</body>
</html>
